refactor: move permissions modal to features roles
